+++ integtest/spec/single_book_spec.rb
@@ -8,12 +8,6 @@ RSpec.describe 'building a single book' do
   end
 
   ##
-  # Build a path to a file in the destination.
-  def dest_file(file)
-    File.expand_path(file, @dest)
-  end
-
-  ##
   # Write a source file.
   def write_source(source_path, text)
     path = source_file source_path
@@ -74,38 +68,6 @@ RSpec.describe 'building a single book' do
     end
   end
 
-  ##
-  # Include in a context named after a page to add an assertion that the page
-  # was created and to create some variables extracted from the page's contents
-  #
-  # body - the non-template body of the page
-  # title - the title of the page
-  shared_context 'page body' do
-    let(:file) do |example|
-      dest_file(example.example_group.description)
-    end
-    it 'is created' do
-      expect(file).to file_exist
-    end
-    let(:body) do
-      return nil unless File.exist? file
-
-      File.open(dest_file(file), 'r:UTF-8') do |f|
-        f.read
-         .sub(/.+<!-- start body -->/m, '')
-         .sub(/<!-- end body -->.+/m, '')
-      end
-    end
-    let(:title) do
-      return nil unless body
-
-      m = body.match(%r{<h1 class="title"><a id=".+"></a>([^<]+)(<a.+?)?</h1>})
-      raise "Can't find title in #{body}" unless m
-
-      m[1]
-    end
-  end
-
   HEADER = <<~ASCIIDOC
     = Title
 
@@ -125,15 +87,13 @@ RSpec.describe 'building a single book' do
       end
       include_context 'convert'
 
-      context 'index.html' do
-        include_context 'page body'
+      page_context 'index.html' do
         it 'has the right title' do
           expect(title).to eq('Title')
         end
       end
-      context 'chapter.html' do
-        include_context 'page body'
-        it 'has the right tile' do
+      page_context 'chapter.html' do
+        it 'has the right title' do
           expect(title).to eq('Chapter')
         end
       end
@@ -168,8 +128,7 @@ RSpec.describe 'building a single book' do
     end
     include_context 'convert'
 
-    context 'chapter.html' do
-      include_context 'page body'
+    page_context 'chapter.html' do
       it 'contains the index text' do
         expect(body).to include('I include "included"')
       end
@@ -192,8 +151,7 @@ RSpec.describe 'building a single book' do
     it 'copies the warning image' do
       expect(dest_file('images/icons/warning.png')).to file_exist
     end
-    context 'chapter.html' do
-      include_context 'page body'
+    page_context 'chapter.html' do
       it 'includes the warning image' do
         expect(body).to include(
           '<img alt="Warning" src="images/icons/warning.png" />'
